Updates starter to 2.0.0
    Major Changes:
    * Updates React and ReactDOM to v16
    * Replaces Mocha with Jest
    * Adds `babel-polyfill` and updates Babel config
    * Colocates tests with components

    Minor Changes:
    * Updates all other dependencies to latest
    * Allows absolute import paths
    * Adds new test scripts
    * Adds test coverage
