+++ .babelrc
@@ -1,3 +1,3 @@
 {
-  "presets": ["es2015", "react"]
+  "presets": ["env", "stage-0", "react"]
 }
