+++ .gitignore
@@ -2,3 +2,4 @@ node_modules/
 *npm-debug.log
 *.DS_Store
 dist/bundle.js
+coverage
