+++ __tests__/cssStub.js
@@ -1,0 +1,3 @@
+// This is a stub file to keep Jest from yelling
+// about the CSS module imports in our components.
+export default {};
