+++ server.js
@@ -9,8 +9,8 @@ app.get('/', function(request, response) {
   response.sendFile(__dirname + '/dist/index.html');
 });
 
-app.listen(PORT, error => {
+app.listen(PORT, error => (
   error
-  ? console.error(error)
-  : console.info(`Listening on port ${PORT}. Visit http://localhost:${PORT}/ in your browser.`)
-});
+    ? console.error(error)
+    : console.info(`Listening on port ${PORT}. Visit http://localhost:${PORT}/ in your browser.`)
+));
